enemies now always roll for wealth drops, regardless of the xp limit. this isn't a problem since the consumables are short lived anyway.

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/Mob.java
@@ -995,19 +995,7 @@ public abstract class Mob extends Char {
 	}
 	
 	public void rollToDropLoot(){
-		if (Dungeon.hero.lvl > maxLvl + 2) return;
 		boolean dropped = false;
-
-		MasterThievesArmband.StolenTracker stolen = buff(MasterThievesArmband.StolenTracker.class);
-		if (stolen == null || !stolen.itemWasStolen()) {
-			if (Random.Float() < lootChance()) {
-				Item loot = createLoot();
-				if (loot != null) {
-					Dungeon.level.drop(loot, pos).sprite.drop();
-					dropped=true;
-				}
-			}
-		}
 		
 		//ring of wealth logic
 		if (Ring.getBuffedBonus(Dungeon.hero, RingOfWealth.Wealth.class) > 0) {
@@ -1022,17 +1010,30 @@ public abstract class Mob extends Char {
 			}
 		}
 		
-		//lucky enchant logic
-		if (buff(Lucky.LuckProc.class) != null){
-			Dungeon.level.drop(buff(Lucky.LuckProc.class).genLoot(), pos).sprite.drop();
-			Lucky.showFlare(sprite);
-			dropped=true;
-		}
-
-		//soul eater talent
-		if (buff(SoulMark.class) != null &&
-				Random.Int(10) < Dungeon.hero.pointsInTalent(Talent.SOUL_EATER)){
-			Talent.onFoodEaten(Dungeon.hero, 0, null);
+		if (Dungeon.hero.lvl <= maxLvl + 2) {
+			MasterThievesArmband.StolenTracker stolen = buff(MasterThievesArmband.StolenTracker.class);
+			if (stolen == null || !stolen.itemWasStolen()) {
+				if (Random.Float() < lootChance()) {
+					Item loot = createLoot();
+					if (loot != null) {
+						Dungeon.level.drop(loot, pos).sprite.drop();
+						dropped=true;
+					}
+				}
+			}
+			
+			//lucky enchant logic
+			if (buff(Lucky.LuckProc.class) != null){
+				Dungeon.level.drop(buff(Lucky.LuckProc.class).genLoot(), pos).sprite.drop();
+				Lucky.showFlare(sprite);
+				dropped=true;
+			}
+			
+			//soul eater talent
+			if (buff(SoulMark.class) != null &&
+					Random.Int(10) < Dungeon.hero.pointsInTalent(Talent.SOUL_EATER)){
+				Talent.onFoodEaten(Dungeon.hero, 0, null);
+			}
 		}
 		
 		Heap heap = Dungeon.level.heaps.get(pos);
